Allows taking pictures over advanced consoles or looking up/down, and fixes some multi-z photography issues. (#91632)
## About The Pull Request

I wanted to allow you to take pictures over advanced camera consoles for
heist planning style point reasons, and parity with the AI, and in the
process found a small pile of bugs.
So, this aims to fix a pile of those, and slightly clean up the code in
the process.

The biggest part of this change involves making the multi-z part of
photography code properly account for turfs below open or glass turfs,
rather than trying to skip to the bottom. This needs to be done because
objects can exist above those open turfs, most commonly
lattices/catwalks, but also nograv stuff.

...Parallax is its whole own other issue with how photography works
right now, so uh. Yeah, no.

Sadly I couldn't get this to work with regular camera consoles, and as I
found it to be an entirely different can of worms that would only
slightly touch this code I decided that would be best explored in a
separate pr.
## Why It's Good For The Game

I think it's cute to be able to take a picture of your screen and show
it off to people, maybe even combine it with a detective board for a
cool little planning thing. And the AI can take pictures over the
cameranet so, y'know.
Doing so while looking up/down was a side-effect, but really I think
it's fine to let you photograph it if you can see it.

Less multi-z jank is good :+1:

### Actual Pictures:
<details>
  <summary>Unfold</summary>
  

![image](https://github.com/user-attachments/assets/322ba400-e72e-4e0e-b974-6ba60372103f)

![image](https://github.com/user-attachments/assets/aaa0ffd4-7313-44ea-bc73-ac55b4c2a51b)

![image](https://github.com/user-attachments/assets/f5647b02-60c4-4d46-a9a7-6f1541550e1d)

![image](https://github.com/user-attachments/assets/49498845-9fdc-4907-b660-96989c0b28fb)

![image](https://github.com/user-attachments/assets/d8c19118-84f2-4ab6-b8c0-0d5427ed9262)

![image](https://github.com/user-attachments/assets/1ade6619-4d7f-42ce-8063-041e8fe58dfb)

![image](https://github.com/user-attachments/assets/b5d14478-4ce0-4f5c-a9bf-1ef927103d85)

</details>

## Changelog
:cl:
add: You can take pictures over advanced camera consoles.
add: You can take pictures while looking up or down.
fix: AIs taking pictures in multi-z shows the objects of your floor and
the ones below, instead of just the ones below.
fix: Humans taking pictures through holes in multi-z actually shows the
floor below, instead of pitch-black nothingness.
fix: Taking pictures of glass floors in multi-z actually shows the floor
below.
/:cl: