Fix runtime when grabbing/helping someone with no suit on (#91721)
## About The Pull Request

Evidently this checks if any item you're wearing that covers your hands
are bloody, and if they are, the blood will pass on. But though it
checks hands cover, it always refers to suit, for some reason.

This is technically my code

Bonus fix: Fixes an inverted check in `attack_effects`