You can see inside closets and interact with yourself and others inside them (#92029)
## About The Pull Request




https://github.com/user-attachments/assets/65a03f4a-0e44-4bce-b603-a019d5a3a36e

I want you to be able to pick up items and stuff in here but it seemed
annoying so I'm letting Melbert figure that out

Closes #92006

I didn't make it work with client log ins and log outs and all that such
and such fyi

## Why It's Good For The Game

Doesn't really make sense that you can't heal yourself in closets when
you're still able to use items and such in there. You're in the closet.

## Changelog
:cl:
add: You can now see inside closets you're in, and interact with
yourself inside.
/:cl: