HoP traitors may get an objective to kill the Captain and steal their ID (#92153)
## About The Pull Request

HoP traitors have a chance to get an objective to "replace the captain"
requiring they assassinate the captain and steal their (or any) gold id.

It can only be rolled if there's a captain (obviously) 

## Why It's Good For The Game

Figured we could give certain jobs fun unique objectives which lean into
some common sci fi tropes, like "untrustworthy second in command". In
this case, it'd be pretty fitting for a hop to have enough and try to
take over the station themself.

However I kept the flavor pretty broad - for some they might take the
objective as "I'm becoming the new captain", for others it might be "I'm
jealous of their cool ID card".

## Changelog

:cl: Melbert
add: HoP traitors can get an objective to "replace the captain" -
requiring they assassinate the captain and steal their id.
/:cl: