Removes the magical healing anti-stun inspiration banners from Catwalk and replace them with their mundane variants (#92422)
## About The Pull Request
These inspiration banners are removed on all the other maps a long long
time ago but they were put on Catwalk. They let you use them in hand to
heal brute, burn, and stuns. As I understand it they aren't meant to be
available to crewmembers outside of admin events and gimmicks.

## Why It's Good For The Game
Fixes an oversight

## Changelog
:cl: Bisar
fix: Correct a few banners on Catwalk Station to be mundane banners, and
not the magic barriers that heal you and remove your stuns.
/:cl: